perrone/requirements.txt
asgiref==3.10.0
blinker==1.9.0
click==8.2.1
Django==5.2.8
django-jazzmin==3.0.1
djangorestframework==3.16.1
Flask==3.1.2
Flask-SQLAlchemy==3.0.5
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.43
sqlparse==0.5.3
typing_extensions==4.15.0
Werkzeug==3.1.3
